python_frontend_desktop/requirements_modern.txt
requests==2.31.0
python-dotenv==1.0.0
plyer==2.1.0
customtkinter==5.2.2
pillow==10.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23